+++ example/lib/ui/views/counter_view/counter_view_model.dart
@@ -1,11 +1,13 @@
 import 'package:easy_mvvm/easy_mvvm.dart';
 
 class CounterViewModel extends EasyViewModel {
+  static const String counterProperty = 'counter';
+
   int _counter = 0;
   int get counter => _counter;
 
   void incrementCounter() {
     _counter++;
-    updateView();
+    notifyPropertyChange(counterProperty);
   }
 }