+++ lib/src/custom_page_route_builder.dart
@@ -14,58 +14,59 @@ class CustomPageRouteBuilder extends PageRouteBuilder {
     required RouteTransition routerTransition,
     required Widget child,
     Duration? transitionDuration,
-    RouteSettings? settings,
+    super.settings,
     Color? backgroundColor,
-  }): super(
-    settings: settings,
-    transitionDuration: transitionDuration ?? const Duration(milliseconds: 300),
-    reverseTransitionDuration: transitionDuration ?? const Duration(milliseconds: 300),
-    pageBuilder: (_, __, ___) => child,
-    transitionsBuilder: (context, animation, secondaryAnimation, child) {
-      switch (routerTransition) {
-        case RouteTransition.none:
-          break;
-        case RouteTransition.appDefault:
-          break;
-        case RouteTransition.fadeScale:
-          return FadeScaleTransition(
-            animation: animation,
-            child: child,
-          );
-        case RouteTransition.fadeThrough:
-          return FadeThroughTransition(
-            animation: animation,
-            secondaryAnimation: secondaryAnimation,
-            fillColor: backgroundColor,
-            child: child,
-          );
-        case RouteTransition.scaledSharedAxis:
-          return SharedAxisTransition(
-            animation: animation,
-            secondaryAnimation: secondaryAnimation,
-            transitionType: SharedAxisTransitionType.scaled,
-            fillColor: backgroundColor,
-            child: child,
-          );
-        case RouteTransition.verticalSharedAxis:
-          return SharedAxisTransition(
-            animation: animation,
-            secondaryAnimation: secondaryAnimation,
-            transitionType: SharedAxisTransitionType.vertical,
-            fillColor: backgroundColor,
-            child: child,
-          );
-        case RouteTransition.horizontalSharedAxis:
-          return SharedAxisTransition(
-            animation: animation,
-            secondaryAnimation: secondaryAnimation,
-            transitionType: SharedAxisTransitionType.horizontal,
-            fillColor: backgroundColor,
-            child: child,
-          );
-      }
+  }) : super(
+          transitionDuration:
+              transitionDuration ?? const Duration(milliseconds: 300),
+          reverseTransitionDuration:
+              transitionDuration ?? const Duration(milliseconds: 300),
+          pageBuilder: (_, __, ___) => child,
+          transitionsBuilder: (context, animation, secondaryAnimation, child) {
+            switch (routerTransition) {
+              case RouteTransition.none:
+                break;
+              case RouteTransition.appDefault:
+                break;
+              case RouteTransition.fadeScale:
+                return FadeScaleTransition(
+                  animation: animation,
+                  child: child,
+                );
+              case RouteTransition.fadeThrough:
+                return FadeThroughTransition(
+                  animation: animation,
+                  secondaryAnimation: secondaryAnimation,
+                  fillColor: backgroundColor,
+                  child: child,
+                );
+              case RouteTransition.scaledSharedAxis:
+                return SharedAxisTransition(
+                  animation: animation,
+                  secondaryAnimation: secondaryAnimation,
+                  transitionType: SharedAxisTransitionType.scaled,
+                  fillColor: backgroundColor,
+                  child: child,
+                );
+              case RouteTransition.verticalSharedAxis:
+                return SharedAxisTransition(
+                  animation: animation,
+                  secondaryAnimation: secondaryAnimation,
+                  transitionType: SharedAxisTransitionType.vertical,
+                  fillColor: backgroundColor,
+                  child: child,
+                );
+              case RouteTransition.horizontalSharedAxis:
+                return SharedAxisTransition(
+                  animation: animation,
+                  secondaryAnimation: secondaryAnimation,
+                  transitionType: SharedAxisTransitionType.horizontal,
+                  fillColor: backgroundColor,
+                  child: child,
+                );
+            }
 
-      return child;
-    },
-  );
+            return child;
+          },
+        );
 }